matamataeraakuria

+++ base/event_store/store/sys_event_repeat_guard.cpp
@@ -63,7 +63,7 @@ bool SysEventRepeatGuard::IsTimeRangeMatched(const std::string& file)
 {
     struct stat fileInfo;
     stat(file.c_str(), &fileInfo);
-    uint64_t modiTime = fileInfo.st_mtime;
+    uint64_t modiTime = static_cast<uint64_t> (fileInfo.st_mtime);
     return modiTime >= GetMinValidTime();
 }
